+++ crates/runner/src/runtime.rs
@@ -2,7 +2,7 @@ use frame_support::{runtime, traits::FindAuthor, weights::constants::WEIGHT_REF_
 use pallet_revive::AccountId32Mapper;
 use polkadot_sdk::*;
 use polkadot_sdk::{
-    polkadot_sdk_frame::{log, runtime::prelude::*},
+    polkadot_sdk_frame::runtime::prelude::*,
     sp_runtime::{AccountId32, Perbill},
 };
 
@@ -72,6 +72,7 @@ parameter_types! {
 
 #[derive_impl(pallet_revive::config_preludes::TestDefaultConfig)]
 impl pallet_revive::Config for Runtime {
+    type Balance = Balance;
     type Time = Timestamp;
     type Currency = Balances;
     type DepositPerByte = DepositPerByte;