+++ CHANGELOG.md
@@ -4,17 +4,24 @@
 
 This is a development pre-release.
 
-Supported `polkadot-sdk` rev: `2509.0.0`
+Supported `polkadot-sdk` rev: `unstable2507`
 
 ### Added
 - The comprehensive revive compiler book documentation page: https://paritytech.github.io/revive/
+- Support for solc v0.8.33.
+- Support for the `clz` Yul builtin.
 
 ### Changed
 - Instruct the LLVM backend and linker to `--relax` (may lead to smaller contract code size).
 - Standard JSON mode: Don't forward EVM bytecode related output selections to solc.
+- The supported `polkadot-sdk` release is `unstable2507`.
+- The `INVALID` opcode and OOB memory accesses now consume all remaining gas.
+- Emit the `call_evm` and `delegate_call_evm` syscalls for contract calls.
 
 ### Fixed:
 - The missing `STOP` instruction at the end of `code` blocks.
+- The missing bounds check in the internal sbrk implementation.
+- The call gas is no longer ignored.
 
 ## v0.5.0